Almost lost
Ugh

+++ libby/libmacho-1.0/segment.h
@@ -24,26 +24,26 @@
 #include <libcrippy-1.0/libcrippy.h>
 
 typedef struct macho_segment_cmd_t {
-	uint32_t cmd;
-	uint32_t cmdsize;
+	uint64_t cmd;
+	uint64_t cmdsize;
 	char segname[16];
-	uint32_t vmaddr;
-	uint32_t vmsize;
-	uint32_t fileoff;
-	uint32_t filesize;
-	uint32_t maxprot;
-	uint32_t initprot;
-	uint32_t nsects;
-	uint32_t flags;
+	uint64_t vmaddr;
+	uint64_t vmsize;
+	uint64_t fileoff;
+	uint64_t filesize;
+	uint64_t maxprot;
+	uint64_t initprot;
+	uint64_t nsects;
+	uint64_t flags;
 } macho_segment_cmd_t;
 
 typedef struct macho_segment_t {
 	char* name;
-	uint32_t size;
-	uint32_t offset;
-	uint32_t address;
+	uint64_t size;
+	uint64_t offset;
+	uint64_t address;
 	unsigned char* data;
-	uint32_t section_count;
+	uint64_t section_count;
 	macho_section_t** sections;
 	macho_segment_cmd_t* command;
 } macho_segment_t;
@@ -52,7 +52,7 @@ typedef struct macho_segment_t {
  * Mach-O Segment Functions
  */
 macho_segment_t* macho_segment_create();
-macho_segment_t* macho_segment_load(unsigned char* data, uint32_t offset);
+macho_segment_t* macho_segment_load(unsigned char* data, uint64_t offset);
 macho_section_t* macho_segment_get_section(macho_segment_t* segment, const char* section);
 void macho_segment_debug(macho_segment_t* segment);
 void macho_segment_free(macho_segment_t* segment);
@@ -61,7 +61,7 @@ void macho_segment_free(macho_segment_t* segment);
  * Mach-O Segment Info Functions
  */
 macho_segment_cmd_t* macho_segment_cmd_create();
-macho_segment_cmd_t* macho_segment_cmd_load(unsigned char* data, uint32_t offset);
+macho_segment_cmd_t* macho_segment_cmd_load(unsigned char* data, uint64_t offset);
 void macho_segment_cmd_debug(macho_segment_cmd_t* cmd);
 void macho_segment_cmd_free(macho_segment_cmd_t* cmd);
